Extracting the history management into its own class
